2023-12-25 04:21 UTC-0500 Ron Pinkas <ronpinkas/AT/gmail/com>   * utils/misc/hbrc.c     unsigned long long -> ULONGLONG
  * source/compiler/hbstripl.c
    * Move HB_SYMBOL_UNUSED() (executable context) below declaration

  * make_bc.bat
    + Added auto detection with auto setup of \BCC55 as SET_BCC_55
    + Added auto detection with auto setup of C:\win_flex_bison-2.5.5 as
    + Added: IF "%CC%"=="bcc32c" SET SUB_DIR=b32c

  * source/rdd/usrrdd/rdds/adordd/make_b32.bat
  * source/rdd/usrrdd/rdds/adordd/makefile.bc
  * tests/dll_b32.mak
  * tests/dll_b32.bat
    * modify hard coded b32 to %SUB_DIR% to differentiate between bcc32 (b32) and bcc32c (b32c)

  * winmake/common3.mak
    * Modified hard coded bison to $(BISON) to allow support for win_bison.exe
  * winmake/makefile.bc
    * Modified OPT_FLAGS of bcc32c to use same flags as bcc32

+++ utils/misc/hbrc.c
@@ -67,6 +67,8 @@
 #include <windows.h>
 #include <time.h>
 #include <stdio.h>
+
+#include "hbdefs.h"
 #include "hbver.h"
 #include "hbverbld.h"
 
@@ -353,7 +355,7 @@ int main( int argc, char * argv[] )
          char *      aMo[]     = { "January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December" };
          char        szParam[ 265 ];
 
-         unsigned long long llvalue = HB_VER_CVSID; 
+         ULONGLONG llvalue = HB_VER_CVSID; 
          unsigned long dw1 = llvalue >> 32; // Get the first DWORD (the higher 32 bits) 
          unsigned long dw2 = llvalue & 0xFFFFFFFF; // Get the second DWORD (the lower 32 bits)
